# Conflicts:
#	kgms/src/main/java/com/plantdata/kgcloud/domain/dw/service/impl/DWServiceImpl.java
#	kgms/src/main/java/com/plantdata/kgcloud/domain/dw/service/impl/PreBuilderServiceImpl.java
#	kgms/src/main/java/com/plantdata/kgcloud/domain/dw/util/PaserYaml2SchemaUtil.java

+++ kgms/src/main/java/com/plantdata/kgcloud/domain/dw/service/impl/GraphMapServiceImpl.java
@@ -299,6 +299,8 @@ public class GraphMapServiceImpl implements GraphMapService {
             if(attrMapList != null && !attrMapList.isEmpty()){
                 deleteConcept(attrMapList);
             }
+
+            return ;
         }
 
         Map<Integer,AttributeDefinitionRsp> existAttrIds = schemaRsp.getAttrs().stream().collect(Collectors.toMap(AttributeDefinitionRsp::getId, Function.identity()));
@@ -328,7 +330,7 @@ public class GraphMapServiceImpl implements GraphMapService {
 
                     List<String> relationAttrNames = attr.getExtraInfos().stream().map(AttrExtraRsp::getName).collect(Collectors.toList());
 
-                    List<DWGraphMapRelationAttr> relationAttrList = graphMapRelationAttrRepository.findAll(Example.of(DWGraphMapRelationAttr.builder().kgName(kgName).attrId(attr.getId()).build()));
+                    List<DWGraphMapRelationAttr> relationAttrList = graphMapRelationAttrRepository.findAll(Example.of(DWGraphMapRelationAttr.builder().kgName(kgName).modelAttrId(graphMap.getModelAttrId()).build()));
 
                     //没有边属性订阅，跳过
                     if(relationAttrList == null || relationAttrList.isEmpty()){
@@ -340,11 +342,15 @@ public class GraphMapServiceImpl implements GraphMapService {
                     //图谱中该属性的边属性不存在，删除订阅记录
                     for(DWGraphMapRelationAttr relationAttr : relationAttrList){
                         if(!relationAttrNames.contains(relationAttr.getName())){
-                            graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            if(graphMapRelationAttrRepository.existsById(relationAttr.getId())){
+                                graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            }
                         }
 
                         if(existRelationAttList.contains(relationAttr.getName()+relationAttr.getModelId()+relationAttr.getModelAttrId())){
-                            graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            if(graphMapRelationAttrRepository.existsById(relationAttr.getId())) {
+                                graphMapRelationAttrRepository.deleteById(relationAttr.getId());
+                            }
                         }else{
                             existRelationAttList.add(relationAttr.getName()+relationAttr.getModelId()+relationAttr.getModelAttrId());
                         }
